build: always call vcpkg_install.bat from cmake
Instead of checking if VCPKG_DIR exists first, always call
vcpkg_install.bat, which will create a new vcpkg clone as well as
perform any necessary actions on an existing clone, and will do nothing
if nothing needs to be done.

Remove the "Initializing vcpkg..." message from cmake and change the
"installing libraries" message from the cmd script to say that it may
take a while.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>