build: support VCPKG_ROOT env in vcpkg_install.bat
If %VCPKG_ROOT% environment variable is set, check that the parent
directory exists and throw an error if it does not.

If %VCPKG_ROOT% is not set, set it to %cwd%\vcpkg, which was the
previous behavior.

Change all references to %cwd%\vcpkg to %VCPKG_ROOT%.

Change the documentation in the top comment to refer to %VCPKG_ROOT%.

This allows a developer to maintain his or her own vcpkg clone and not
have to unnecessarily repeat the very time-consuming process of
compiling the dependencies.

Set ENV{VCPKG_ROOT} to the VCPKG_DIR default in CMakeLists.txt, because
the cmake code does not support this yet, this is introduced in a
subsequent commit.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>