build: add build_option() to cmake/Utilities.cmake
Add the build_option() function which is similar to the cmake option()
command, but allows adding alias variables and alias environment
variables.

The option and all aliases are set to the first found value by listed
sequence in the option and all aliases, followed by the default if
provided, or 'OFF' if not provided, with cache variables having
precedence over normal variables.

In the case of environment variables or aliases, if the type is BOOL and
the found/default value is OFF/FALSE unset the environment variable.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>