build: clean up NO_GETTEXT cmake support
Make NO_GETTEXT a build option aliased to ENV{NO_GETTEXT} to pass to
vcpkg_install.bat to skip building gettext.

Turn it off by default when vcpkg is in-use because it's a very long
compile.

Do not build translations if NO_GETTEXT is set.

Add a note about NO_GETTEXT to the docs at the top of CMakeLists.txt.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>